# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bike marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /showrooms/{id}:
    get:
      operationId: getShowroom
      tags: [showrooms]
      summary: Get public showroom profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Showroom profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Showroom"
        "404":
          description: Not found
  /admin/showrooms:
    get:
      operationId: adminListShowrooms
      tags: [admin]
      summary: Admin - list all showrooms
      responses:
        "200":
          description: All showrooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Showroom"
    post:
      operationId: adminCreateShowroom
      tags: [admin]
      summary: Admin - create showroom with its account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShowroomBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Showroom"
        "409":
          description: Username already exists
  /admin/showrooms/{id}:
    put:
      operationId: adminUpdateShowroom
      tags: [admin]
      summary: Admin - update showroom (optionally reset password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShowroomBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Showroom"
    delete:
      operationId: adminDeleteShowroom
      tags: [admin]
      summary: Admin - delete showroom, its account and its bikes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /showroom/bikes:
    get:
      operationId: showroomListBikes
      tags: [showrooms]
      summary: Showroom - list own bikes
      responses:
        "200":
          description: Showroom's bikes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bike"
    post:
      operationId: showroomCreateBike
      tags: [showrooms]
      summary: Showroom - create a bike listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBikeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bike"
  /showroom/bikes/{id}:
    put:
      operationId: showroomUpdateBike
      tags: [showrooms]
      summary: Showroom - update own bike listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShowroomBikeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bike"
    delete:
      operationId: showroomDeleteBike
      tags: [showrooms]
      summary: Showroom - delete own bike listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Bike:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
          description: "electric, motorcycle, or bicycle sub-type: mountain, road, hybrid, kids"
        condition:
          type: string
          description: "new, used"
        brand:
          type: string
        phone:
          type: string
        images:
          type: array
          items:
            type: string
        mileage:
          type: number
        engineCapacity:
          type: number
          description: "motorcycle engine capacity in cc"
        province:
          type: string
          description: "one of the 18 Iraqi governorates"
        hasDelivery:
          type: boolean
        hasDocuments:
          type: boolean
          description: "motorcycle has official registration papers"
        status:
          type: string
          description: "active, pending, sold, rejected"
        showroomId:
          type: integer
          nullable: true
        showroom:
          $ref: "#/components/schemas/ShowroomSummary"
          nullable: true
        userId:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        isFavorited:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - price
        - category
        - condition
        - phone
        - status
        - userId
        - createdAt
    CreateBikeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        condition:
          type: string
        brand:
          type: string
        phone:
          type: string
        images:
          type: array
          items:
            type: string
        mileage:
          type: number
        engineCapacity:
          type: number
        province:
          type: string
        hasDelivery:
          type: boolean
        hasDocuments:
          type: boolean
      required:
        - title
        - price
        - category
        - condition
        - phone
        - province
...
    ShowroomSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        googleMapsUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        verified:
          type: boolean
      required:
        - id
        - name
        - verified
    Showroom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        googleMapsUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        verified:
          type: boolean
        username:
          type: string
          nullable: true
        bikesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - verified
    CreateShowroomBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
        googleMapsUrl:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - username
        - password
    UpdateShowroomBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
        googleMapsUrl:
          type: string
        phone:
          type: string
        verified:
          type: boolean
        password:
          type: string
          description: "if provided, resets the showroom account password"
    UpdateShowroomBikeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        condition:
          type: string
        brand:
          type: string
        phone:
          type: string
        images:
          type: array
          items:
            type: string
        mileage:
          type: number
        engineCapacity:
          type: number
        province:
          type: string
        hasDelivery:
          type: boolean
        hasDocuments:
          type: boolean
        status:
          type: string
          description: "active or sold"
...